# docs/api-contracts/openapi.inventory-api.yaml
openapi: 3.0.3
info:
  title: Inventory Mobile API
  description: |
    Este archivo OpenAPI define el contrato REST esperado para la aplicacion
    movil de inventario.

    Direccion planificada:
    - La implementacion prevista del backend es ASP.NET Core Web API.
    - Flutter consumira esta API usando Dio / HttpClient.
    - SQL Server sera la capa de persistencia planificada detras del backend.
    - Docker y Docker Compose se usaran para la infraestructura del backend.

    Estado:
    - Este contrato no implica que el backend, SQL Server o Docker Compose ya
      esten implementados o configurados.
    - Los endpoints, esquemas y ejemplos describen la forma esperada de la
      integracion cuando el backend este disponible.

    Relacion con Firebase:
    - Firebase no es la capa principal de persistencia del sistema.
    - Firebase sigue siendo relevante para Firebase Cloud Messaging (FCM).
    - Firebase Storage puede usarse opcionalmente para imagenes de productos si
      el equipo decide adoptarlo.
  version: 1.0.0
servers:
  - url: https://api.inventory.example.com
    description: Production-like environment placeholder.
  - url: http://localhost:8080
    description: Local development environment.
...
security:
  - bearerAuth: []
paths:
...
  /notification-tokens:
    post:
      tags:
        - NotificationTokens
      summary: Register or upsert a device token
      description: |
        Registers a push notification token for the authenticated user. If
        the same token is already recorded, the existing record is updated
        and associated with the authenticated user instead of duplicated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationTokenCreateRequest'
            example:
              token: fcm_device_token_admin_demo
              platform: android
      responses:
        '201':
          description: Token registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationToken'
              example:
                id: 17862eaf-2ed2-4550-aa7f-e031915d23e6
                token: fcm_device_token_admin_demo
                platform: android
                createdAt: '2026-06-02T20:00:00Z'
                updatedAt: null
        '200':
          description: Existing token updated and associated with the authenticated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationToken'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Token could not be registered because of an unresolved persistence conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /notification-tokens/{tokenId}:
    parameters:
      - in: path
        name: tokenId
        required: true
        schema:
          type: string
          format: uuid
        description: Identifier of the notification token.
    delete:
      tags:
        - NotificationTokens
      summary: Remove a device token
      description: |
        Removes a push notification token. Used on logout or when the
        client detects that the token has been invalidated by the platform.
      responses:
        '204':
          description: Token removed.
        '400':
          description: tokenId is not a valid GUID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Token not found or not owned by the authenticated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PlatformType:
      type: string
      enum:
        - android
        - ios
        - web
      description: Device platform reported for a push notification token.
...
    NotificationToken:
      type: object
      required:
        - id
        - token
        - platform
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: 17862eaf-2ed2-4550-aa7f-e031915d23e6
        token:
          type: string
          maxLength: 500
          example: fcm_device_token_admin_demo
        platform:
          $ref: '#/components/schemas/PlatformType'
        createdAt:
          type: string
          format: date-time
          example: '2026-06-02T20:00:00Z'
        updatedAt:
          type: string
          format: date-time
          nullable: true
          example: '2026-06-02T20:00:00Z'
...
    NotificationTokenCreateRequest:
      type: object
      required:
        - token
        - platform
      properties:
        token:
          type: string
          minLength: 1
          maxLength: 500
          example: fcm_device_token_admin_demo
        platform:
          $ref: '#/components/schemas/PlatformType'
...
    FieldError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: name
        message:
          type: string
          example: Name is required.
    ErrorBody:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code.
          example: validation_error
        message:
          type: string
          description: Human-readable error message.
          example: The request contains invalid fields.
        details:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        requestId:
          type: string
          example: req_123
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      example:
        error:
          code: validation_error
          message: The request contains invalid fields.
          details:
            - field: name
              message: Name is required.
          requestId: req_123
...